User account

+++ server/models/User.js
@@ -4,7 +4,7 @@ module.exports = (sequelize, DataTypes) => {
   var User = sequelize.define('User', {
     firstName: DataTypes.STRING,
     lastName: DataTypes.STRING,
-    saveGoal: DataTypes.INTEGER,
+    saveGoal: DataTypes.DECIMAL(12,2),
     email: DataTypes.STRING,
     password: DataTypes.STRING,
     googleId: DataTypes.STRING,
